+++ packages/intents-sdk/src/bridges/omni-bridge/omni-bridge-constants.ts
@@ -1,7 +1,7 @@
 export const NEAR_NATIVE_ASSET_ID = "nep141:wrap.near";
 export const OMNI_BRIDGE_CONTRACT = "omni.bridge.near";
-export const MIN_ALLOWED_STORAGE_BALANCE_FOR_INTENTS_NEAR =
-	500000000000000000000000n; // 0.5 NEAR
+export const MIN_STORAGE_BALANCE_FOR_INTENTS_NEAR = 500000000000000000000000n; // 0.5 NEAR
 
 // This was chosen empirically and may not be optimal, it includes 10% margin.
 export const MIN_GAS_AMOUNT = "37400000000000"; // 37.4 tgas
+export const INTENTS_STORAGE_BALANCE_CACHE_KEY = "INTENTS_STORAGE_BALANCE";